Do not cache fetch calls inside `"use cache"` implicitly
If a `fetch` call inside of `"use cache"` has not specified `cache:
'force-cache'` or `revalidate`, we should not cache it implicitly in the
fetch cache. Instead, it's only cached as part of the surrounding cached
function.

Those `fetch` calls that do specify `cache: 'force-cache'` or
`revalidate` are still cached in the fetch cache. They're like an inner
`"use cache"`. They also don't inherit the tags of the outer `"use
cache"`.

partially reverts #71793 and #72075
supersedes #72090