Refactor implicitTags
This change is part of a larger set of refactors meant to eventually remove the requesStore that is currently scoped around the entire renderToHTML... function call. Ideally workUnitStore would be scoped to specific render tasks and in this change I am disentangling one dependency on this outer store by directly passing implicitTags into the prerender implementation to remove one dependency on this outer store.