Refactor prerender phase assignment
When prerendering there really shouldn't be a requestStore at all. I will be removing it but in this change I am specifically removing the assignment of the render phase in the prerender function. This isn't needed because the phase is render by default anyway and all relevant rendering must be scoped inside a prerenderStore which precludes a requestStore anyway. This is part of a larger refactor to remove requestStore from shadowing prerenders.

This change should not alter any program semantics