Refactor: move url to app context to avoid passing requestStore
requestStore is set on the app context but this store doens't make sense to scope for all renders, in particular prerenders so we need to extract the necessary information that is actually globally available and carry that around on the app context instead. This change updates app context to carry the url object and it updates reads of the url to use app context instead of requestStore. Now nothing is reading requestStore from the context so I removed it.

This update is part of a larger refactor to eliminate the requestStore that shadows prerendering. It should not alter any program semantics.