refactor(turbopack-core): Migrate more resolve options structs to ResolvedVc (#72200)
A hand refactor of all the remaining struct fields from `Vc` to `ResolvedVc`, to unblock local tasks.

This is similar to #71948, as there's a focus on resolve options.

Note: "resolve" in this context means two different things. `turbopack-resolve` is about resolving module locations on disk. `ResolvedVc` is about resolving "cells" representing incremental computations.

Closes PACK-3369