Refactor prerender cookies (#72211)
the prerender pathway had carryover logic from a prior refactor that
checks if any cookies were modified during the prerender before encoding
a redirect. This is non necessary because prerenders by definition have
no mutable cookies. This is part of a larger refactor to eliminate the
requestStore that scopes renderToHTML... including when prerendering.
This shouldn't be necessary and if it is we have not factored the
program sufficiently.

Stacked on #72207