Snowflake backend for devlow (#72162)
This adds a backend to devlow for Vercel’s HTTP data api for sending
values to Snowflake. It requires the gateway’s batch uri endpoint,
schema id, and topic name.

It collects the metrics and sends them as a batch, each with relevant
metadata similar to the metadata collected by the Datadog backend.

Test Plan:

- [x] Manual test
- [x] Unit test


Closes PACK-3363