Turbopack: Ignore `Unknown module type` for NFT (#71713)
Part of NFT

~~Dynamic requests such as `require("./"+ something)` often match too
many files. If we cannot process a matched file with the Turbopack
rules, then it's really an unreachable case anyway. This removes the
error in that case.~~

This doesn't change the behavior of regular bundling, but makes
`primary_modules_raw_iter` not throw the error.

- [x] Cleanup API: add the boolean into `.process()`? Or should this be
passed differently?
- [ ] ~~Should this be a warning instead for user code?~~