Remove the node-file-trace crate (#70998)
### What?

This is a 2nd prefactor for adding node-file-trace into turbopack. When implementing this, we have a few options namely:

- a) expose turbotrace (node-file-trace-rs) as a napi library and try to insert it into the current codebase
- b) expose turbotrace and try to reuse the turbo backend
- c) move turobtrace into the rust process and do it in parallel with everything else

We chose c

We have already ripped out the js code path, this PR nukes the now-orphaned node-file-trace rust binary and napi bindings.

### Why?

The crate is unmaintained and probably will not be used in its current form.