ensure fetch is patched before preloading entries (#72363)
If a user creates a custom fetch implementation in module scope, the
logic to preload entries during `next start` will result in fetch not
being patched before the userland implementation. As a result, it won't
opt into any caching heuristics from our patched fetch.

Fixes #71881 (in addition to the other PR in this stack)