Do not log full error when source map can not be resolved
When a source map can't be resolved, logging the full error including
the stack trace unnecessarily clutters the terminal. For Next.js users,
this is a non-actionable exception. The only viable action is to report
it as an issue. Next.js maintainers can use `__NEXT_VERBOSE_LOGGING`
when investigating those issues.