Include bound args param in server reference information byte arg mask
Considering the bound args param in the server reference information
byte is required for a client-side optimization of omitting unused
arguments to work properly when a server action or `"use cache"`
function uses closed-over variables.

The server reference information byte is not used yet, so this PR should
not change any observable behavior.