[Segment Prefetch] Move access token to route tree
Since prefetches include only static data, and are generated during
the server prerender, we can actually lift all the tokens to the route
tree prefetch. It's only once we start supporting dynamic conditional
routes that the tokens will need to provided by the layout.

This is nice because it means we can parallelize the segment prefetches
after all.