Extract shared logic to separate functions
In the Segment Cache implementation, I won't be using the main
fetchFromServerResponse function because the protocol for per-segment
prefetches is different. But there's a lot of logic (e.g. setting
headers, decoding the Flight stream, validating the response) that can
be shared. So I've extracted the shared bits into separate functions.